+++ src/app/ap_app.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import "./ap_app.css"; // Import global styles
-import Navbar from "../../components/devices/navbar/ap_Navbar";
+import Navbar from "../components/devicesPage/navbar/ap_Navbar";
 //import { Sidebar } from '../../components/devices/Sidebar';
-import DeviceManagement from "../../components/devices/main/ap_DeviceManagement";
+import DeviceManagement from "../components/devicesPage/main/ap_DeviceManagement";
 
 export default function App() {
   return (
