+++ src/components/devicesPage/additional/ap_button.jsx
