+++ src/components/devicesPage/additional/ap_card.jsx
