+++ src/components/devicesPage/ap_tailwind.config.jsx
