+++ src/components/devicesPage/main/ap_DeviceManagement.module.css
