+++ src/components/devicesPage/navbar/ap_Navbar.jsx
