+++ src/components/devicesPage/navbar/ap_Navbar.module.css
