+++ src/components/devicesPage/navbar/ap_Weather.jsx
