+++ src/components/devicesPage/navbar/ap_Weather.module.css
