+++ src/pages/index.jsx
@@ -1,5 +1,5 @@
-import React from 'react';
-import App from '../app/profiles/ap_app';
+import React from "react";
+import App from "../app/ap_app";
 
 export default function Home() {
   return (
